use workaround version of mintablefuntoken contract
Getting this frontend to work with the borsh-serialized version of the
`mint` endpoint was not working, so a workaround was added in
43454e84f6a7eced62467c59bb0fe5d4e7f57ee0

I've deployed a workaround version of the contract to testnet, so it can
be used here. The name of this contract is meant to be ridiculous to
remind anyone paying attention that this is NOT the long-term version of
this contract interface.