WIP: use generic connector & already-initialized ERC20
The generic token connector can work with any ERC20, but a token factory
on the NEAR side needs to deploy a sub-contract for each new ERC20. This
ERC20 has been initialized in this way already. Once I get it working
with this token, I can figure out the interface for initializing others,
and add UI for it.