feat: add UI for selecting a featured ERC20
The signed-in UI now shows a select box with a list of featured ERC20
tokens. Once you select one, the URL updates and the form is shown to
send one.

This also includes some usability-focused tweaks to helper functions:

* `urlParams.clear` clears the `?` from the URL when clearing last param
* `fill` keeps a cache and doesn't re-paint needlessly. This means that
  when you select an ERC20 to send but then cancel, the select dropdown
  still has the same item selected.

Note that filling in your own erc20 address in the URL is supported,
but the UI itself does not expose this functionality.