feat: almost send a Bridged NEP21 back to Ethereum
This adds (disabled!) UI for sending Bridged NEP21 tokens from NEAR back
to Ethereum, and largely copies the testing code from the `from-near.js`
file in the `github.com/near/rainbow-bridge` repository, but adapts it
to a browser-based environment.

Why disabled? Because implementing this change requires completion of
modifications to near-api-js, which are not quite functional. However,
this is a large change, and to enable parallel efforts on this codebase,
we need to merge this now.

This required a large refactor of the `transfers.js` code:

* split transfers.js into multiple files
* transfers.js exports main interface for initiating and managing
  transfers
* localStorage has been moved to transfers/storage, with exported
  functions slightly higher-level than raw localStorage wrapping
* all natural-erc20-to-nep21 transfer logic has been moved to own file
* some (safe?) console errors were coming from html/send-erc20. While
  not completely related to the other changes here, these errors made it
  hard to troubleshoot other problems during this refactor, and so have
  been fixed here.