explore: init-send-inspect explicit NEAR Wallet loop
This explores some potential interfaces we could add to near-api-js to
make it easier to code an "init-send-inspect" loop when an app expects
the "send" step to cause a redirect to NEAR Wallet.

1. init: initialize transaction locally, storing transaction hash in
   local storage (in the code here, this hash gets stored within the
   `transfer` object).
2. send: the app needs to set URL Params prior to sending the
   transaction, to set things up for the "inspect" step.
3. inspect: the app has figured out that it just sent a transaction
   based on URL Params, and can easily fetch the same info about this
   transaction that would have been returned by a near-api-js contract
   call that required no redirect.

Note that, as coded in this particular app, this code would work whether
or not the contract call causes a redirect.