explore: use 'listener' for NEAR Wallet redirects
This writes some non-functional code to explore a potential update to
near-api-js:

* Make contract calls accept an object as a second argument, rather than
  positional arguments for `gas` and `deposit`
* You can add a `listener` key to this object, and pass it a function.
  Whether or not the contract call results in a redirect to NEAR Wallet,
  this function will be the first thing called after the call completes.

How would this work, if the contract call results in a redirect?
near-api-js would need to serialize (a reference to) the function and
save it in localStorage. Then, as soon as `WalletConnection` is
instantiated, near-api-js would check if such a function is saved, and
invoke it.

Is it possible to serialize and save a function like this? I sort of
doubt it!