explore: deal with NEAR Wallet redirect using queue
This adds some non-functional code to explore a possible way to deal
with NEAR Wallet redirects. This approach ONLY works if the contract
call does, in fact, cause a redirect.

The basic idea is to add the transaction to an `inFlightTransactions`
queue rather than sending it directly, which allows adding some metadata
so that the app can identify what transaction this is for upon initial
page load.

In this particular app, that metadata is used to find the correct
`transfer` and determine which step of the transfer it's at, so it can
call the correct next function.

Without significant rework, this approach feels doomed to lead to
hard-to-understand code, with related functionality spread throughout
the codebase, and exported constants & functions that didn't previously
need to be exported.