refactor: disable sending Bridged NEP21 for now
We need to lay some groundwork for parallel efforts on this repository.
Given that this change restructures the code signficantly, we need to
make sure it's merged before multiple people start working on separate
parts of the codebase at once.

This allows the code to be merged safely, while avoiding any errors due
to the current implementation not actually working.