feat: add dumbest possible syncing
This syncs all data between all users! Obviously a terrible syncing
approach, but the easiest proof-of-concept to make quickly.

Figuring out a good way to add authentication to PouchDB/CouchDB is
actually fairly BYO and more complicated than I hoped. One COULD use
keys prefixed with a given user's eth & NEAR addresses, then filter for
these records for a given user. But this obviously makes it easy for a
malicious user to add noise, spam, and potentially worse attacks that
would show up for other users.

The [database per user][1] approach seems close to what we would want,
but it wasn't obvious how to pull this off with Cloudant, the backend
being used here. And besides, what we really want is blockchain-based
verification. This would probably require a small service to sit on the
server between the PouchDB frontend connections and CloudDB.

  [1]: https://github.com/pouchdb-community/pouchdb-authentication/blob/master/docs/recipes.md#couch_peruser-configuration