feat: "new transfer" form
When you click "Begin new transfer" from the list of transfers, show the
"swappable authenticated accounts" UI that was shown on the landing
before you started your first transfer.

This required extracting "swappable authenticated accounts" to a new
authenticated-accounts.html file, and modifying its behavior to make it
still work when added to multiple places on the page. This, in turn,
prompted some cleanup to authNear & authEthereum, moving rendering logic
to `renderers` pushed in by authenticated-accounts.html.

Finally, an optimization snuck in: an `init` function is again exported
from `domHelpers`, and this time the only logic in it will accept all
forms submitted with `action="get"` and avoid page refreshes for them.