fix: [workaround] withdraw by switching auth
The `withdraw` transaction does not (currently) require an attached
deposit. This means it can be called without causing a redirect, if the
user is signed in against the specific BridgeToken contract.

When the user first signs into NEAR in this app, they sign in against
the BridgeTokenFactory.

This change checks if they're signed in against the given BridgeToken,
and if not, clears their NEAR keys and then signs them in against the
BridgeToken. It can then call `withdraw` and check the result of this
transaction without problem.

This fix revealed further problems, some of which have also been
addressed here.

Additionally, this no longer requires a special in-progress branch of
near-api-js, so the dependency has been changed to the latest branch.