refactor: nep21→nep141; mv deploy & getMetadata to lib
This started off as an effort to remove the
`window.nearFungibleTokenFactory` from `authNear` so that the extracted
library would only depend on the presence of `window.nearConnection`,
and ended with moving an enormous amount of code and renaming NEP21 to
NEP141.

The shape of the main exports from the connector client library
(`nep141~erc20`) has changed, and the documentation in that folder has
been updated accordingly.