fix: update token balances without page refresh
The list of ERC20 tokens and the list of associated bridged NEP141
variants was only rendered once, at page load. This meant that if you
started a transfer of a certain number of tokens then started a new
transfer, you would see your old balance in the token-select modals.

This fixes some caching issues in the getMetadata helpers to efficiently
re-render the tokens on every `render` call.